refactor: 重构任务与日志状态管理，统一使用int8枚举类型
本次提交完成了多项核心重构：
1. 新增LogStatus枚举统一管理任务状态，替换原字符串状态
2. 将所有任务、日志、配置的状态字段从string改为int8类型，提升性能和一致性
3. 重构算法、任务、字典等模型的状态枚举值映射，修正状态序号偏移
4. 优化数据库查询条件，移除冗余的deleted过滤条件
5. 统一模型基类使用，简化实体定义
6. 修复文件ID日志打印的类型错误
7. 调整JSON字段类型，适配数据库存储格式

+++ dehaze-go/internal/model/vo/algorithm_version.go
@@ -13,7 +13,7 @@ type AlgorithmVersionVO struct {
 	// 变更日志
 	ChangeLog *string `json:"changeLog"`
 	// 状态
-	Status *int `json:"status"`
+	Status *int8 `json:"status"`
 	// 是否当前活跃版本
 	IsActive *bool `json:"isActive"`
 	// 模型文件ID
